+++ src/step/DirectoryStepTest.ts
@@ -3,17 +3,23 @@ import { SsgContextImpl } from "../SsgContextImpl"
 import { SsgContext } from "../SsgContext"
 import { describe, expect, test } from "@javarome/testscript"
 import { SsgFile } from "../util"
+import path from "path"
 
 describe("DirectoryStep", () => {
 
   test("processes subdirs", async () => {
-    const config = {outDir: "out"}
     const name = "test directory step"
     const step = new class extends DirectoryStep {
       protected async processDirs(_context: SsgContext, _dirNames: string[], outputFile: SsgFile): Promise<void> {
         return outputFile.write()
       }
-    }(["src/"], [], "test/test.html", config, name)
+    }({
+      rootDirs: ["src/"], getOutputPath(_context: SsgContext): string {
+        return path.join("out", this.templateFileName)
+      },
+      excludedDirs: [], templateFileName: "test/test.html"
+
+    }, name)
     expect(step.name).toBe(name)
     const context = new SsgContextImpl("fr")
     const result = await step.execute(context)