+++ src/step/content/ContentStepConfig.ts
@@ -1,7 +1,8 @@
 import { SsgContext } from "../../SsgContext"
 import { ReplaceCommand } from "./replace"
+import { SsgConfig } from "../../SsgConfig"
 
-export type ContentStepConfig<C extends SsgContext = SsgContext> = {
+export interface ContentStepConfig<C extends SsgContext = SsgContext> extends SsgConfig {
   /**
    * The glob roots of contents to process.
    */
@@ -11,10 +12,4 @@ export type ContentStepConfig<C extends SsgContext = SsgContext> = {
    * The replacements to process on contents.
    */
   replacements: ReplaceCommand<C>[],
-
-  /**
-   * @param context
-   * @return the file where to output.
-   */
-  getOutputFile(context: C): string
 }